CHE-1000: create project structure
