+++ assembly/assembly-ide-war/src/main/webapp/WEB-INF/classes/codenvy/che.properties
@@ -65,9 +65,9 @@ schedule.core_pool_size=10
 ### for the machines can come from DockerHub or a private Docker registry.
 machine.docker.registry=${CHE_REGISTRY_HOST}:5000
 docker.registry.auth.url=http://${CHE_REGISTRY_HOST}:5000
-docker.registry.auth.username=user1
-docker.registry.auth.password=pass
-docker.registry.auth.email=user1@email.com
+docker.registry.auth.username=NULL
+docker.registry.auth.password=NULL
+docker.registry.auth.email=NULL
 docker.connection.tcp.connection_timeout_ms=600000
 docker.connection.tcp.read_timeout_ms=600000
 
@@ -80,14 +80,13 @@ machine.server.ext.archive=${che.home}/lib/ws-agent.zip
 
 # The location of the Web Socket terminal used within the browser.
 # This is copied into the machine and run from within it.
-machine.server.terminal.archive=${che.home}/lib/terminal
+machine.server.terminal.archive=${che.home}/lib/linux_amd64/terminal
 
 # The machine's log files are stored here
 machine.logs.location=${che.logs.dir}/machine/logs
 
 # Size of the machine by default.  What is used if RAM parameter not provided by user or API.
 machine.default_mem_size_mb=1024
-machine.supported_docker_version=1.6.0,1.6.1,1.6.2,1.7.1,1.8.1
 
 # When the workspace master launches a new workspace, Che performs checks of the internal Web 
 # services. When Che gets a valid response, we know that the workspace agent is ready for use.
@@ -109,6 +108,9 @@ machine.docker.che_api.endpoint=http://che-host:${SERVER_PORT}/ide/api
 # If this is true, then we always pull an image from a registry even if we have an image cached
 # locally. If false, Docker only pulls image if it does not exist locally.
 machine.docker.pull_image=true
+
+# If true, then all docker machines will start in privilege mode.
+machine.docker.privilege_mode=false
 
 # If the browser clients that are accessing Che are remote AND the configuration of Docker is an
 # internal IP address or using Unix sockets, then remote browser clients will not be able to connect
@@ -134,3 +136,7 @@ org.everrest.asynchronous.job.timeout=10
 org.everrest.asynchronous.cache.size=1024
 # Path to asynchronous service
 org.everrest.asynchronous.service.path=/async/
+
+machine.ssh.connection_timeout_ms=3000
+machine.ssh.server.terminal.path_to_archive.linux_amd64=${che.home}/lib/linux_amd64/terminal
+machine.ssh.server.terminal.path_to_archive.linux_arm7=${che.home}/lib/linux_arm7/terminal
