+++ assembly/assembly-machine-server/pom.xml
@@ -14,17 +14,17 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>che-assembly-parent</artifactId>
+        <artifactId>artik-ide-assembly-parent</artifactId>
         <groupId>org.eclipse.che</groupId>
-        <version>4.0.0-SNAPSHOT</version>
+        <version>4.2.0-RC1-SNAPSHOT</version>
     </parent>
-    <artifactId>assembly-machine-server</artifactId>
+    <artifactId>artik-ide-assembly-machine-server</artifactId>
     <packaging>pom</packaging>
-    <name>Che Assembly Machine Plugins :: Server</name>
+    <name>Artik IDE :: Assembly Machine Plugins :: Server</name>
     <dependencies>
         <dependency>
             <groupId>org.eclipse.che</groupId>
-            <artifactId>assembly-machine-war</artifactId>
+            <artifactId>artik-ide-assembly-machine-war</artifactId>
             <type>war</type>
         </dependency>
         <dependency>
