+++ assembly/assembly-machine-server/src/assembly/assembly.xml
@@ -31,7 +31,7 @@
             <outputDirectory>webapps</outputDirectory>
             <outputFileNameMapping>ide.war</outputFileNameMapping>
             <includes>
-                <include>org.eclipse.che:assembly-machine-war</include>
+                <include>org.eclipse.che:artik-ide-assembly-machine-war</include>
             </includes>
         </dependencySet>
     </dependencySets>
