+++ assembly/assembly-machine-war/src/main/webapp/WEB-INF/classes/codenvy/che-machine-configuration.properties
@@ -12,9 +12,6 @@
 # vfs
 user.local.db=NULL
 
-# storage path for models from local dao implementations
-che.conf.storage=/local-storage
-
 # everrest.async
 # Disable asynchronous mechanism that is embedded in everrest. Codenvy has own mechanism for that.
 org.everrest.asynchronous=false
