+++ assembly/assembly-main/pom.xml
@@ -14,22 +14,22 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>che-assembly-parent</artifactId>
+        <artifactId>artik-ide-assembly-parent</artifactId>
         <groupId>org.eclipse.che</groupId>
-        <version>4.0.0-GA-SNAPSHOT</version>
+        <version>4.2.0-RC1-SNAPSHOT</version>
     </parent>
-    <artifactId>assembly-main</artifactId>
+    <artifactId>artik-ide-assembly-main</artifactId>
     <packaging>pom</packaging>
-    <name>Che IDE :: Assemblies Tomcat</name>
+    <name>Artik IDE :: Assemblies Tomcat</name>
     <dependencies>
         <dependency>
             <groupId>org.eclipse.che</groupId>
-            <artifactId>assembly-ide-war</artifactId>
+            <artifactId>artik-ide-assembly-ide-war</artifactId>
             <type>war</type>
         </dependency>
         <dependency>
             <groupId>org.eclipse.che</groupId>
-            <artifactId>assembly-machine-server</artifactId>
+            <artifactId>artik-ide-assembly-machine-server</artifactId>
             <type>zip</type>
         </dependency>
         <dependency>
@@ -61,6 +61,13 @@
             <groupId>org.eclipse.che.lib</groupId>
             <artifactId>che-websocket-terminal</artifactId>
             <type>zip</type>
+            <classifier>linux_amd64</classifier>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.che.lib</groupId>
+            <artifactId>che-websocket-terminal</artifactId>
+            <type>zip</type>
+            <classifier>linux_arm7</classifier>
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.plugin</groupId>
