+++ assembly/assembly-main/src/assembly/assembly.xml
@@ -25,7 +25,7 @@
             <outputDirectory>tomcat/webapps</outputDirectory>
             <outputFileNameMapping>ide.war</outputFileNameMapping>
             <includes>
-                <include>org.eclipse.che:assembly-ide-war</include>
+                <include>org.eclipse.che:artik-ide-assembly-ide-war</include>
             </includes>
         </dependencySet>
         <dependencySet>
@@ -52,15 +52,28 @@
             <outputDirectory>lib</outputDirectory>
             <outputFileNameMapping>ws-agent.zip</outputFileNameMapping>
             <includes>
-                <include>org.eclipse.che:assembly-machine-server</include>
+                <include>org.eclipse.che:artik-ide-assembly-machine-server</include>
             </includes>
         </dependencySet>
         <dependencySet>
             <useProjectArtifact>false</useProjectArtifact>
             <unpack>true</unpack>
-            <outputDirectory>lib</outputDirectory>
+            <outputDirectory>lib/linux_amd64</outputDirectory>
             <includes>
-                <include>org.eclipse.che.lib:che-websocket-terminal</include>
+                <include>org.eclipse.che.lib:che-websocket-terminal:zip:linux_amd64</include>
+            </includes>
+            <unpackOptions>
+                <excludes>
+                    <exclude>META-INF/**</exclude>
+                </excludes>
+            </unpackOptions>
+        </dependencySet>
+        <dependencySet>
+            <useProjectArtifact>false</useProjectArtifact>
+            <unpack>true</unpack>
+            <outputDirectory>lib/linux_arm7</outputDirectory>
+            <includes>
+                <include>org.eclipse.che.lib:che-websocket-terminal:zip:linux_arm7</include>
             </includes>
             <unpackOptions>
                 <excludes>
@@ -95,6 +108,32 @@
                 <include>ch.qos.logback:logback-core</include>
             </includes>
         </dependencySet>
+        <dependencySet>
+            <useProjectArtifact>false</useProjectArtifact>
+            <unpack>true</unpack>
+            <outputDirectory>stacks</outputDirectory>
+            <includes>
+                <include>org.eclipse.che.core:che-core-ide-stacks</include>
+            </includes>
+            <unpackOptions>
+                <excludes>
+                    <exclude>META-INF/**</exclude>
+                </excludes>
+            </unpackOptions>
+        </dependencySet>
+        <dependencySet>
+            <useProjectArtifact>false</useProjectArtifact>
+            <unpack>true</unpack>
+            <outputDirectory>templates</outputDirectory>
+            <includes>
+                <include>org.eclipse.che.core:che-core-ide-templates</include>
+            </includes>
+            <unpackOptions>
+                <excludes>
+                    <exclude>META-INF/**</exclude>
+                </excludes>
+            </unpackOptions>
+        </dependencySet>
     </dependencySets>
     <fileSets>
         <fileSet>
@@ -113,14 +152,6 @@
             <outputDirectory>plugins</outputDirectory>
         </fileSet>
         <fileSet>
-            <directory>${project.basedir}/src/assembly/stacks</directory>
-            <outputDirectory>stacks</outputDirectory>
-        </fileSet>
-        <fileSet>
-            <directory>${project.basedir}/src/assembly/templates</directory>
-            <outputDirectory>templates</outputDirectory>
-        </fileSet>
-        <fileSet>
             <directory>${project.basedir}/src/assembly/webapps</directory>
             <outputDirectory>tomcat/webapps</outputDirectory>
         </fileSet>
@@ -137,29 +168,29 @@
            </includes>
         </fileSet>
         <fileSet>
-            <directory>${project.parent.basedir}/assembly-machine-war</directory>
-            <outputDirectory>sdk/assembly-machine-war</outputDirectory>
+            <directory>${project.parent.basedir}/artik-ide-assembly-machine-war</directory>
+            <outputDirectory>sdk/artik-ide-assembly-machine-war</outputDirectory>
             <excludes>
             <exclude>**/target/**</exclude>
             </excludes>
         </fileSet>
         <fileSet>
-            <directory>${project.parent.basedir}/assembly-machine-server</directory>
-            <outputDirectory>sdk/assembly-machine-server</outputDirectory>
+            <directory>${project.parent.basedir}/artik-ide-assembly-machine-server</directory>
+            <outputDirectory>sdk/artik-ide-assembly-machine-server</outputDirectory>
             <excludes>
             <exclude>**/target/**</exclude>
             </excludes>
         </fileSet>
         <fileSet>
-            <directory>${project.parent.basedir}/assembly-ide-war</directory>
-            <outputDirectory>sdk/assembly-ide-war</outputDirectory>
+            <directory>${project.parent.basedir}/artik-ide-assembly-ide-war</directory>
+            <outputDirectory>sdk/artik-ide-assembly-ide-war</outputDirectory>
             <excludes>
             <exclude>**/target/**</exclude>
             </excludes>
         </fileSet>
         <fileSet>
-            <directory>${project.parent.basedir}/assembly-main</directory>
-            <outputDirectory>sdk/assembly-main</outputDirectory>
+            <directory>${project.parent.basedir}/artik-ide-assembly-main</directory>
+            <outputDirectory>sdk/artik-ide-assembly-main</outputDirectory>
             <excludes>
             <exclude>**/target/**</exclude>
             </excludes>
