--- che-plugin-artik-ide/src/main/java/org/eclipse/che/plugin/artik/ide/ArtikResources.java
@@ -1,26 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012-2016 Codenvy, S.A.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Codenvy, S.A. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.che.plugin.artik.ide;
-
-import com.google.gwt.core.client.GWT;
-import com.google.gwt.resources.client.ClientBundle;
-
-import org.vectomatic.dom.svg.ui.SVGResource;
-
-/**
- * @author Vitalii Parfonov
- */
-public interface ArtikResources extends ClientBundle {
-    ArtikResources INSTANCE = GWT.create(ArtikResources.class);
-
-    @Source("svg/artik.svg")
-    SVGResource artikIcon();
-}
