--- che-plugin-artik-ide/src/main/java/org/eclipse/che/plugin/artik/ide/manager/ArtikDeviceManager.java
@@ -1,24 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012-2016 Codenvy, S.A.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Codenvy, S.A. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.che.plugin.artik.ide.manager;
-
-import javax.inject.Inject;
-
-/**
- *
- */
-public class ArtikDeviceManager {
-
-    @Inject
-    public ArtikDeviceManager(){
-
-    }
-}
