--- che-plugin-artik-ide/src/main/java/org/eclipse/che/plugin/artik/ide/manager/RegisterDeviceViewImpl.ui.xml
@@ -1,80 +1,0 @@
-<!--
-
-    Copyright (c) 2012-2016 Codenvy, S.A.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Codenvy, S.A. - initial API and implementation
-
--->
-<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
-<ui:UiBinder xmlns:ui='urn:ui:com.google.gwt.uibinder'
-             xmlns:g='urn:import:com.google.gwt.user.client.ui'>
-    <ui:with field='localizationConstant' type='org.eclipse.che.plugin.artik.ide.ArtikLocalizationConstant'/>
-
-    <ui:style>
-        .emptyBorder {
-            margin: 6px;
-        }
-
-        .inputField {
-            color: #dbdbdb;
-            border: 1px solid #191c1e;
-            border-radius: 1px;
-            box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.19) inset, 0px 1px 0px 0px rgba(223, 223, 223, 0.4);
-            font-size: 12px;
-            padding-left: 8px;
-            height: 20px;
-        }
-
-        .label {
-            line-height: 25px;
-        }
-
-        .floatLeft {
-            float: left;
-        }
-
-        .floatRight {
-            float: right;
-        }
-
-        .errorHint {
-            color: red;
-            height: 12px;
-            margin: 10px;
-        }
-    </ui:style>
-
-    <g:DockLayoutPanel unit="PX" width="500px" height="160px" addStyleNames="{style.emptyBorder}">
-        <g:north size="40">
-            <g:FlowPanel>
-                <g:Label text="{localizationConstant.ipAddress}"
-                         addStyleNames="{style.label} {style.floatLeft}"/>
-                <g:TextBox width="400px" ui:field="ipAddress" debugId="registerArtikDevice-ip"
-                           addStyleNames="{style.inputField} {style.floatRight}"/>
-            </g:FlowPanel>
-        </g:north>
-        <g:center>
-            <g:FlowPanel>
-                <g:Label text="{localizationConstant.registerArtikDeviceLogin}"
-                         addStyleNames="{style.label} {style.floatLeft}"/>
-                <g:TextBox width="400px" ui:field="login" debugId="registerArtikDevice-login"
-                           addStyleNames="{style.inputField} {style.floatRight}"/>
-            </g:FlowPanel>
-        </g:center>
-        <g:south size="40">
-            <g:FlowPanel>
-                <g:Label text="{localizationConstant.registerArtikDevicePassword}"
-                         addStyleNames="{style.label} {style.floatLeft}"/>
-                <g:TextBox width="400px" ui:field="password" debugId="registerArtikDevice-password"
-                           addStyleNames="{style.inputField} {style.floatRight}"/>
-            </g:FlowPanel>
-        </g:south>
-
-
-    </g:DockLayoutPanel>
-</ui:UiBinder>
