+++ plugins/plugin-artik-ide/src/main/java/org/eclipse/che/plugin/artik/ide/inject/ArtikGinModule.java
@@ -1,0 +1,28 @@
+/*******************************************************************************
+ * Copyright (c) 2012-2016 Codenvy, S.A.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *   Codenvy, S.A. - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.che.plugin.artik.ide.inject;
+
+import com.google.gwt.inject.client.AbstractGinModule;
+import org.eclipse.che.ide.api.extension.ExtensionGinModule;
+
+/**
+ * @author Vitalii Parfonov
+ */
+@ExtensionGinModule
+public class ArtikGinModule extends AbstractGinModule {
+
+    /** {@inheritDoc} */
+    @Override
+    protected void configure() {
+    }
+
+
+}
