+++ plugins/plugin-artik-ide/src/main/resources/org/eclipse/che/plugin/artik/Artik.gwt.xml
