+++ plugins/plugin-artik-server/pom.xml
@@ -14,19 +14,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>artik-ide-parent</artifactId>
-        <groupId>org.eclipse.che</groupId>
+        <artifactId>artik-ide-plugin-parent</artifactId>
+        <groupId>org.eclipse.che.plugin</groupId>
         <version>4.2.0-RC1-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
     </parent>
-    <artifactId>artik-ide-assembly-parent</artifactId>
-    <version>4.2.0-RC1-SNAPSHOT</version>
-    <packaging>pom</packaging>
-    <name>Artik IDE :: Assemblies parent</name>
-    <modules>
-        <module>assembly-machine-war</module>
-        <module>assembly-machine-server</module>
-        <module>assembly-ide-war</module>
-        <module>assembly-main</module>
-    </modules>
+    <artifactId>artik-ide-plugin-server</artifactId>
+    <packaging>jar</packaging>
+    <name>Artik IDE :: Plugin :: Samsung Artik :: Server Component</name>
+    <dependencies />
 </project>
