+++ plugins/plugin-artik-shared/src/main/java/org/eclipse/che/plugin/artik/shared/dto/ArtikDeviceInfo.java
